refactor: change the structure of the project and rename some components.
- Use the name of the application (PhotoFx) for the main components.
- Create a compose package to have the common components and modifiers.
- Create a ui package for each one of the screens.
- Isolate MainUiState and EffectConfig common interfaces.
