+++ composeApp/src/androidMain/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:label="Photo FX"
         android:theme="@android:style/Theme.Material.NoActionBar">
         <activity
-            android:name=".AppActivity"
+            android:name=".PhotoFxActivity"
             android:configChanges="orientation|screenSize|screenLayout|keyboardHidden"
             android:launchMode="singleInstance"
             android:windowSoftInputMode="adjustPan"
