--- composeApp/src/androidMain/kotlin/com/mmartosdev/photofx/App.android.kt
@@ -1,28 +1,0 @@
-package com.mmartosdev.photofx
-
-import android.app.Application
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-
-class AndroidApp : Application() {
-    companion object {
-        lateinit var INSTANCE: AndroidApp
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        INSTANCE = this
-    }
-}
-
-class AppActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            App(onCloseClicked = { finish() })
-        }
-    }
-}
