+++ composeApp/src/iosMain/kotlin/main.kt
@@ -1,9 +1,9 @@
 import androidx.compose.ui.window.ComposeUIViewController
-import com.mmartosdev.photofx.App
+import com.mmartosdev.photofx.PhotoFxApplication
 import platform.UIKit.UIViewController
 import kotlin.system.exitProcess
 
 fun MainViewController(): UIViewController =
     ComposeUIViewController {
-        App(onCloseClicked = { exitProcess(0) })
+        PhotoFxApplication(onCloseClicked = { exitProcess(0) })
     }
