+++ composeApp/src/jsMain/kotlin/main.kt
@@ -1,13 +1,13 @@
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.CanvasBasedWindow
-import com.mmartosdev.photofx.App
+import com.mmartosdev.photofx.PhotoFxApplication
 import org.jetbrains.skiko.wasm.onWasmReady
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
     onWasmReady {
         CanvasBasedWindow("Photo FX") {
-            App(onCloseClicked = null)
+            PhotoFxApplication(onCloseClicked = null)
         }
     }
 }
