+++ composeApp/src/jvmMain/kotlin/main.kt
@@ -2,7 +2,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
-import com.mmartosdev.photofx.App
+import com.mmartosdev.photofx.PhotoFxApplication
 import java.awt.Dimension
 
 fun main() = application {
@@ -12,8 +12,6 @@ fun main() = application {
         onCloseRequest = ::exitApplication,
     ) {
         window.minimumSize = Dimension(350, 600)
-        App(
-            onCloseClicked = ::exitApplication,
-        )
+        PhotoFxApplication(onCloseClicked = ::exitApplication)
     }
 }
